SSE-2575: Create a router and controller for clients
Split the services controller into two to ensure they each have a single responsibility.
Mount the clients routes from the services routes as they are dependent on the parent service.
Pull the `serviceId` from the request context and check it's not null where needed.
